(PC-27843) refactor(ConfirmChangeEmail): Extract content to deprecated file
