+++ src/features/profile/pages/ConfirmChangeEmail/ConfirmChangeEmail.tsx
@@ -1,56 +1,11 @@
-import React, { useCallback, useEffect } from 'react'
-import { NativeStackScreenProps } from 'react-native-screens/native-stack'
+import React from 'react'
 
-import { ApiError } from 'api/ApiError'
-import { navigateToHome, navigateToHomeConfig } from 'features/navigation/helpers'
-import { RootStackParamList } from 'features/navigation/RootNavigator/types'
-import { useConfirmChangeEmailMutationV1 } from 'features/profile/helpers/useConfirmChangeEmailMutationV1'
-import { useEmailUpdateStatus } from 'features/profile/helpers/useEmailUpdateStatus'
-import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
-import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
-import { SNACK_BAR_TIME_OUT, useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
-import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
+import { ConfirmChangeEmailContentDeprecated } from 'features/profile/pages/ConfirmChangeEmail/ConfirmChangeEmailContentDeprecated'
 import { GenericInfoPageWhite } from 'ui/pages/GenericInfoPageWhite'
 import { BicolorPhonePending } from 'ui/svg/icons/BicolorPhonePending'
-import { Invalidate } from 'ui/svg/icons/Invalidate'
 import { Spacer, Typo } from 'ui/theme'
 
-type ConfirmChangeEmailProps = NativeStackScreenProps<RootStackParamList, 'ConfirmChangeEmail'>
-
-export function ConfirmChangeEmail({ route: { params }, navigation }: ConfirmChangeEmailProps) {
-  const { data: emailUpdateStatus, isLoading: isLoadingEmailUpdateStatus } = useEmailUpdateStatus()
-  const { showErrorSnackBar } = useSnackBarContext()
-  const { mutate, isLoading } = useConfirmChangeEmailMutationV1({
-    onSuccess: () => navigation.navigate('TrackEmailChange'),
-    onError: (error) => {
-      if (error instanceof ApiError && error.statusCode === 401) {
-        navigation.navigate('ChangeEmailExpiredLink')
-        return
-      }
-      showErrorSnackBar({
-        message: 'Désolé, une erreur technique s’est produite. Veuillez réessayer plus tard.',
-        timeout: SNACK_BAR_TIME_OUT,
-      })
-      navigateToHome()
-    },
-  })
-
-  useEffect(() => {
-    if (!isLoadingEmailUpdateStatus) {
-      if (!emailUpdateStatus) {
-        navigateToHome()
-      }
-      if (emailUpdateStatus?.expired) {
-        navigation.navigate('ChangeEmailExpiredLink')
-      }
-    }
-  }, [emailUpdateStatus, isLoadingEmailUpdateStatus, navigation])
-
-  const onConfirmEmail = useCallback(
-    () => mutate({ token: params?.token }),
-    [params?.token, mutate]
-  )
-
+export function ConfirmChangeEmail() {
   return (
     <GenericInfoPageWhite
       icon={BicolorPhonePending}
@@ -59,20 +14,7 @@ export function ConfirmChangeEmail({ route: { params }, navigation }: ConfirmCha
       separateIconFromTitle={false}
       mobileBottomFlex={0.3}>
       <Spacer.Column numberOfSpaces={40} />
-      <ButtonPrimary
-        wording="Confirmer la demande"
-        accessibilityLabel="Confirmer la demande"
-        onPress={onConfirmEmail}
-        disabled={isLoading}
-      />
-      <Spacer.Column numberOfSpaces={4} />
-      <InternalTouchableLink
-        as={ButtonTertiaryBlack}
-        wording="Fermer"
-        navigateTo={navigateToHomeConfig}
-        icon={Invalidate}
-        disabled={isLoading}
-      />
+      <ConfirmChangeEmailContentDeprecated />
     </GenericInfoPageWhite>
   )
 }
